Fix warnings cause by rename
